Fix bubble position on updates
Bubble was not being positioned relative to it's parent, which is what
the centerpoint was set to. This fixes that. It also divides the
accelerometer val by only 10, not 30, so it's easier to see changes.
Gravity ensures there will always be a component that is in the
thousands though, so that will always be wacky.

TODO: I should max it out at the width or height of the parent.